regcomp.c: qr/[\xFF]/di doesn't have runtime dependencies
Prior to this commit, a class containing U+FF, LATIN SMALL LETTER Y WITH
DIAERESIS, generated an ANYOFD regnode because it thought that what
matched depended on the UTF-8ness of the target string.  But it doesn't.
No bugs were introduced because when ANYOFD is encountered the code
looks at some flags to determine what sorts of dependencies to further
look for, and the flags remained clear.  But ANYOFD is less desirable
than plain ANYOF, because it adds extra branches to execute.

Tests for this fix will be added in a future commit.