Add regnode EXACTFUP, for problematic
If a non-UTF-8 pattern contains a MICRO SIGN, this special node is now
created.  This character is the only one not needing UTF-8 to represent,
but its fold does need UTF-8, which causes some issues, so it has to be
specially handled.  When matching against a non-UTF-8 target string, the
pattern is effectively folded, but not if the target is UTF-8.  By
creating this node, we can remove the special handling required for the
nodes that don't have a MICRO SIGN, in a future commit.