Improve Configure detection of memmem() [perl #133760].
Linux systems have memmem, but the header prototype is only visible if
the C library, but didn't check if the correct prototype is available.
This patch compiles & runs a test program that will fail if the prototype
is needed but not available.

This does not completely close [perl #133760].  The tests for strlcat()
and strlcpy() may also need to be similarly changed.  Also, this patch
does not change whether _GNU_SOURCE is defined or not.  Presumably that
would be done separately in the linux hints file.