Split t/re/fold_grind.t into multiple test files
This has been a goal for a long time, but I thought it would be a lot of
work, but now have realized that there was a fairly easy simplistic
approach.

The core file is renamed fold_grind.pl.  It formerly had an outer loop
which iterated over the possible character set regex pattern modifiers,
/a, /l, etc that were tested.  Now that loop is just a block and new
wrapper files have been created, one per modifier.  They just pass a
global to the core file that gives which modifier this test file is to
use.  Hence each file corresponds to one iteration of the old outer
loop, splitting the tests up into 6 smaller tests that can run in
parallel.