pp.c: Don't use function call for easy copy
Like the previous commit, this code is adding the UTF-8 for a Greek
character to a string.  It previously used Copy, but this character is
representable as two bytes in both ASCII and EBCDIC UTF-8, the only
character sets that Perl will ever supports, so we can use the
specialized code that is used most everywhere else for two byte UTF-8
characters, avoiding the function overhead, and having to treat this
character as particularly special.