pp.c: Don't assume worst case memory needs
Since 5.28, there has been a function that will calculate the expansion
of a string when converted into UTF-8, using per-word operations.  This
means it runs 8 times faster than doing this count previously would have
taken.

I've come to believe it is better to calculate how much memory we need
than to overallocate based on worst-case scenarios.  This is because in
very large strings, over allocating can lead to unnecessary inefficient
processing.

This commit changes several instances in pp.c where a string needs to be
converted to UTF-8 to not assume the worst case, but instead calculate
what's needed using the faster function.