Improve detection of memrchr, strlcat, and strlcpy.
This is continuation of commit f8d82a1010 addressing [perl #133760].
Linux systems using the musl C library have memmem, memrchr, strlcat, and
strlcpy, but the prototypes are only visible if _GNU_SOURCE is defined.
This patch makes Configure test both whether the prototype is visible
and whether the C symbol is visible.

Still to be done is automatically adding _GNU_SOURCE if the musl library
is being used -- probably in hints/linux.sh.