regcomp.c: Under /l any < 256 char can match any other
The code knew this, but it was adding the ASCII alphabetics to the list
of things that matched in UTF-8 locales.  This is unnecessary, as we've
long had the infrastructure elsewhere to handle all potential mappings
from a Latin1 code point to other Latin1, so we can just rely on it.
And it created complexities for future commits in this series.

The MICRO SIGN is the exception, as it folds to non-Latin1 in UTF-8
locales, and this is the place where the structure exists to handle
that.