t/re/fold_grind.pl: Enhance to deal with Turkic rules
The CaseFolding.txt file has special locale-dependent rules.  This
commit changed fold_grind to notice them, and to generate tests for
the situation we aren't in, which are expected to fail.

Since, as of this commit, the Turkic locale is not recognized, this
commit has the effect of generating tests for the Turkic locale, running
them, and making sure they fail when appropriate.