Bump Devel::PPPort to keep in sync with CPAN
During Perl 5.29.8 release, Devel-PPPort was uploaded
to CPAN but the distro was incomplete as unit tests
were not part of the submitted tarball.

To fix this, a new 3.45 version was published to CPAN.
This commit, keep blead in sync with this change.

(manually cherry picked from commit 80e7c5414423d633f11ec93a7990915e97489502)