PATCH: [perl #134031] BBC breake Math::Clipper
This turns out to be that strtod() is less accurate on this platform
than atof().  I looked at the difference, and a test failed that was
failing in this module was the difference between a 0 and a 1 in the
final digit.

The solution is to forbid in the hints file using strtod