PATCH: [perl #133860] 5.30 regression
These bugs stem from trying to compile a user-defined \p{IsProperty}
before the data for the property is available.  In the past, a bug used
the wrong package for IsProperty, and it wasn't found, so its expansion
was delayed until runtime.  But that bug got fixed, and now it finds the
property and thinks its deliberately empty, at compile time.

This is a change in behavior, even if it is fixing a bug, where the real
problem is unobvious.  The solution adopted in this commit is to defer
all empty properties at pattern compilation time.  If they are still
empty at runtime, that's what the expansion will be.