regcomp.c: Fix -Dr bug
If dumping the program and a single range crosses the border between
being in the bitmap and not, the range must be split at the border
because the output has separate text for things in the bitmap vs. those
not.

I'm not sure that there is a situation where this currently occurs, but
it will so with a future commit