embed.fnc: Change s flag to S
autodoc.pl has the s flag in the '=for apidoc' lines mean something
completely different, but Devel::PPPort combines both these lines and
the contents of embed.fnc into one file, expecting the flags to be
consistent.

This commit changes the 's' in embed.fnc to be 'S', which means the
function is static and has an 'S_' prefix.  Changing embed.fnc keeps the
changes more contained, and lessens the chance that some module will be
adversely affected (I didn't see anything likely in grepping cpan).