embed.fnc: Change x flag to e
autodoc.pl has the x flag in the '=for apidoc' lines mean something
completely different, but Devel::PPPort combines both these lines and
the contents of embed.fnc into one file, expecting the flags to be
consistent.

This commit changes the 'x' in embed.fnc to be 'e', which means to not
export this function.  Changing embed.fnc and makedef.pl keeps the
changes more contained, and lessens the chance that some module will be
adversely affected (I didn't see anything likely in grepping cpan).