embed.fnc: Change M flag to x
The M flag in embed.fnc and the x flag in the '=for apidoc' lines both
mean the same, thing, the entity may change, it is experimental.
Devel::PPPort in places uses the M flag, and in other places, the x.
That will be fixed in a separate commit.

This commit changes the 'M' in embed.fnc to be 'x'.  Changing embed.fnc
and makedef.pl keeps the changes more contained, and lessens the chance
that some module will be adversely affected (I didn't see anything
likely in grepping cpan).