autodoc.pl: Make 's' flag independent of 'n'
The 's' flag is used by autodoc.pl to add a semi-colon to the end of a
usage example.  This commit allows that to happen even if the macro has
parameters.

This allows two macros in scope.c to be properly documented.  Previously
they weren't, and caused warnings in Devel::PPPort as a result.

This commit adds documentation to embed.fnc for the 's' flag, which has
long existed, undocumented.