autodoc.pl: Use embed.fnc entry when available
Having two different pod definitions for the same function leads to
things being out of sync.  This commit ignores the one in the '=for
apidoc' line if there is one in embed.fnc.

Doing so led to a bunch of changes in perlapi, showing that the apidoc
lines were inaccurate.  The ones in embed.fnc must be accurate enough to
get perl to compile.

I added a warning for when there's a redundant entry, and in the next
commit I will remove the many such.