Remove redundant info on =for apidoc lines
This information is already in embed.fnc, and we know it compiles.  Some
of this information is now out-of-date.  Get rid of it.

There was one bit of information that was (apparently) wrong in
embed.fnc.  The apidoc line asked that there be no usage example
generated for newXS.  I added that flag to the embed.fnc entry.