(perl #134125) make no digits after 0x/0b fatal
My original change in 7259f4194 silently adjusted the parse
position to immediately after the 0 in an incomplete hex or binary
literal.  In most cases this leads to a syntax error, but in some
cases, especially with 0x, this can lead to a silent change in
behaviour.

So throw an error instead.