Temporary bump PPPort to 3.54 before next release
This is making porting/cmp_version.t happy for now.
We plan to release Devel::PPPort over the next days/weeks
and this would be required before the next release.

Note that previous Perl version was release without
releasing Devel::PPPort. We probably want to release
a 3.53 version matching it, followed by a 3.54 one.