Begin restore of case sensitive symbols on VMS
There is a long-unmaintained option to build with case sensitive
symbols, which is needed by some extensions that have symbols
that differ only by case and need the Perl core to be built with
the same option.

There will be additional steps necessary, but getting "perl" and
"globals" in the correct case is a start.  We no longer need to
explicitly specify globals for the linker, so don't do that and
thus avoid having to guess which case it wants.