parts/inc/utf8: Fix utf8_to_uvchr_buf for 5.7.0
It turns out that 5.6.1 is more advanced than 5.7.0.  So this commit
uses a further check to decide if it can use a substitute function that
provides decent error checking, or if it has to fall back to one that
doesn't.

(cherry picked from commit 8e3920d656441ef0ba0a5dcef98c1064aab26412)
Signed-off-by: Nicolas R <atoomic@cpan.org>