[perl #132474] Add function for parsing sub signatures
This exposes a GRAMSUBSIGNATURE top-level production from perly.y for
toke.c to consume, which parses a subroutine signature, inside the
parens.

This needed a small change to the existing rules, to pull out a rule
that handles all of the insides of the parens but not the parens
themselves.

(h/t to ilmari for the suggestion)