parts/inc/inctools: Add function to format version
This takes a version number and creates a floating point version with
exactly 7 digits with a decimal point after the 5.  This is used as the
first line in files in parts/todo,base, as well as in the data in
ppport.h.

(cherry picked from commit fad2103a9a2c28ff439b9b7c851a9e09194e91a6)
Signed-off-by: Nicolas R <atoomic@cpan.org>